Add grid search script and test.
Grid search is called like this:
python -m bootstrap.grid -o <path> --gpu 0.5 --cpu 10

This will run training on all available gpus and cpus, with 10 cpus per
training and 0.5 gpus assigned for each training (2 jobs per gpu).

The option file must contain a new option: "gridsearch:"
which contains a list of options to modify, and their
associated values should be a list containing all the values
to test.

See the example file in tests/options-grid.yaml

TODO: analysis at the end. This will need an api to get the best results
for a given run, which exists only as a script for now (compare.py)